// src/share/enums.ts
export const enum ConstraintType {
  TANGENT_TOP = 1,
  TANGENT_RIGHT = 2,
  TANGENT_BOTTOM = 3,
  TANGENT_LEFT = 4,
  TANGENT_FRONT = 5,
  TANGENT_BACK = 6,
  ALIGN_LEFT = 7,
  ALIGN_RIGHT = 8,
  ALIGN_TOP = 9,
  ALIGN_BOTTOM = 10,
  ALIGN_FRONT = 11,
  ALIGN_BACK = 12,
  ALIGN_HORIZONTAL = 13,
  ALIGN_VERTICAL = 14,
}